\name{gamma.ratios}
\docType{data}
\alias{gamma.ratios}
\title{Microarray data for halobacteria exposed to gamma raditaion}
\description{
  A 2400 x 16 matrix of log10 ratios (experiment vs. control) of
  halobacterium exposed to gamma radiation.
}
\usage{gamma.ratios}
\format{numerical matrix with  row (gene) and column
  (condition) names.}
\source{The microarray pipeline at the Institute for Systems Biology.}
\references{
   forthcoming...
}
\keyword{datasets}
