\name{testGaggleUtilPackage}
\alias{testGaggleUtilPackage}
\title{Run unit tests for the GaggleUtil Package}
\description{There are six tests defined: data types supported in the GaggleUtil:
  \enumerate{
  \item test.normalize ()
  \item test.normalize.withNAs ()
  \item test.allPositiveRowNames ()
  \item test.filterVector ()
  \item test.fcor ()
  \item test.sharedSigCondsSearch ()
  }
  For more details, please examine the source: unitTests/gaggleUtilTest.R.

}
\usage{testGaggleUtilPackage ()}
\value{the status of the last test}
\keyword{utilities}

